backup_flask_version/requirements.txt
flask
python-docx
werkzeug
docx2pdf
img2table
pytesseract
opencv-contrib-python-headless==4.9.0.80
Pillow
requests
pywebview
